requirements.txt
flask>=2.0
flask-cors>=3.0
mysql-connector-python>=8.0